+++ integration/logtalk_swi.pl
@@ -1,18 +1,18 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Integration file for SWI-Prolog
-%  Last updated on May 3, 2018
+%  Last updated on February 10, 2021
 %
-%  This file is part of Logtalk <https://logtalk.org/>  
+%  This file is part of Logtalk <https://logtalk.org/>
 %  Copyright 1998-2021 Paulo Moura <pmoura@logtalk.org>
 %  SPDX-License-Identifier: Apache-2.0
 %
 %  Licensed under the Apache License, Version 2.0 (the "License");
 %  you may not use this file except in compliance with the License.
 %  You may obtain a copy of the License at
-%  
+%
 %      http://www.apache.org/licenses/LICENSE-2.0
-%  
+%
 %  Unless required by applicable law or agreed to in writing, software
 %  distributed under the License is distributed on an "AS IS" BASIS,
 %  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -22,11 +22,11 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
-:-	getenv('LOGTALKHOME', LogtalkHome0),
-	prolog_to_os_filename(LogtalkHome, LogtalkHome0),
+:- initialization((
 	% load Logtalk core files
-	atom_concat(LogtalkHome, '/adapters/swi.pl', AdapterFile), consult(AdapterFile),
-	atom_concat(LogtalkHome, '/paths/paths.pl', PathsFile), consult(PathsFile),
-	atom_concat(LogtalkHome, '/integration/logtalk_comp_swi.pl', IntegrationFile), consult(IntegrationFile),
+	load_files('$LOGTALKHOME/adapters/swi.pl', [expand(true)]),
+	load_files('$LOGTALKHOME/paths/paths.pl', [expand(true)]),
+	load_files('$LOGTALKHOME/integration/logtalk_comp_swi.pl', [expand(true)]),
 	% load integration hooks with SWI-Prolog native developer tools
-	atom_concat(LogtalkHome, '/adapters/swihooks.pl', HooksFile), consult(HooksFile).
+	load_files('$LOGTALKHOME/adapters/swihooks.pl', [expand(true)])
+)).